Add packaging for plugins, plus fix for packaging for mythtv

+++ mythplugins/lightpackage.sh
@@ -7,8 +7,8 @@ cd $scriptpath
 sourcedir=$1
 subrelease=$2
 if [[ "$sourcedir" == "" ]] ; then
-    echo Before running this the latest version must be built
-    echo with --prefix=/usr --runprefix=/usr and installed
+    echo Before running this the latest version mythtv and plugins 
+    echo must be built and mythtv lightpackage must be done.
     echo Parameter 1 = install dir
     echo Parameter 2 = subrelease number
     exit 2
@@ -19,21 +19,34 @@ packagever=`git describe|cut -c2-`-$subrelease
 installdir=`readlink -f ../../`
 arch=`dpkg-architecture -q DEB_TARGET_ARCH`
 codename=`lsb_release -c|cut -f 2`
-packagename=mythtv-light_${packagever}_${arch}_$codename
+mythtvpackagename=mythtv-light_${packagever}_${arch}_$codename
+packagename=mythplugins-light_${packagever}_${arch}_$codename
 echo Package $packagename
 if [[ -d $installdir/$packagename ]] ; then
     ls -l $installdir
     echo $installdir/$packagename already exists - run with a subrelease number
     exit 2
 fi
+if [[ ! -d $installdir/$mythtvpackagename ]] ; then
+    echo $installdir/$mythtvpackagename does not exist.
+    echo Build mythtv package first
+    exit 2
+fi
 rm -rf $installdir/$packagename $installdir/$packagename.deb
 #make install INSTALL_ROOT=$installdir/$packagename |& tee $installdir/makeinstall.out
 cp -a "$sourcedir/" "$installdir/$packagename/"
+# Remove mythtv files so that only plugin files remain
+cd $installdir/$packagename
+(cd $installdir/$mythtvpackagename ; find . \( -type f -o -type l \) -print0) | xargs -0 rm -f
+rc=0
+while [[ "$rc" == 0 ]] ; do
+    find . -type d -empty -print0 | xargs -0 rmdir -v || rc=$?
+done
 mkdir -p $installdir/$packagename/DEBIAN
 strip -g -v `find $installdir/$packagename/usr/bin/ -type f -executable`
 strip -g -v `find $installdir/$packagename/usr/lib/ -type f -executable`
 cat >$installdir/$packagename/DEBIAN/control <<FINISH
-Package: mythtv-light
+Package: mythplugins-light
 Version: $packagever
 Section: graphics
 Priority: optional
@@ -41,25 +54,22 @@ Architecture: $arch
 Essential: no
 Installed-Size: `du -B1024 -d0 $installdir/$packagename | cut  -f1`
 Maintainer: Peter Bennett <pgbennett@comcast.net>
-Depends: libavahi-compat-libdnssd1, libqt5widgets5, libqt5script5, libqt5sql5-mysql, libqt5xml5, libqt5network5, libqt5webkit5, libexiv2-13 | libexiv2-14, pciutils
+Depends: mythtv-light
 Conflicts: mythtv-common, mythtv-frontend, mythtv-backend
 Homepage: http://www.mythtv.org
-Description: MythTV Light
- Light weight package that installs MythTV in one package, front end
- and backend. Does not install database or services. This package runs 
- the front end successfully on Raspberry Pi 2 or better. Backend 
- programs are not recommended to be run on a Raspberry Pi.
+Description: MythTV Plugins Light
+ MythTV plugins for the MythTV Light package. MythTV Light must be installed
+ before this package. The following plugins are included for Raspberry Pi:
+  MythArchive 
+  MythBrowser
+  MythGame
+  MythMusic
+  MythNetvision
+  MythNews
+  MythWeather
+  MythZoneMinder
+ The MythGallery plugin is not included as it requires OpenGL
 FINISH
-
-mkdir -p $installdir/$packagename/usr/share/applications/
-cp -f lightpackage/mythtv.desktop $installdir/$packagename/usr/share/applications/
-cp -f lightpackage/mythtv-setup.desktop $installdir/$packagename/usr/share/applications/
-
-mkdir -p $installdir/$packagename/usr/share/pixmaps/
-cp -f lightpackage/mythtv.png $installdir/$packagename/usr/share/pixmaps/
-
-mkdir -p $installdir/$packagename/usr/share/menu/
-cp -f lightpackage/mythtv-frontend $installdir/$packagename/usr/share/menu/
 
 cd $installdir
 fakeroot dpkg-deb --build $packagename
